+++ benches/json_import.rs
@@ -55,8 +55,8 @@ fn json_import_benchmark(c: &mut Criterion) {
                 b.iter(|| {
                     let mut blobs = MemoryBlobStore::<Blake3>::new();
                     let mut importer = JsonObjectImporter::<_, Blake3>::new(&mut blobs, None);
-                    importer.import_blob(blob.clone()).expect("import JSON");
-                    std::hint::black_box(importer.data().len());
+                    let fragment = importer.import_blob(blob.clone()).expect("import JSON");
+                    std::hint::black_box(fragment.facts().len());
                 });
             },
         );
@@ -67,8 +67,8 @@ fn json_import_benchmark(c: &mut Criterion) {
                 b.iter(|| {
                     let mut blobs = MemoryBlobStore::<Blake3>::new();
                     let mut importer = JsonTreeImporter::<_, Blake3>::new(&mut blobs, None);
-                    importer.import_blob(blob.clone()).expect("import JSON");
-                    std::hint::black_box(importer.data().len());
+                    let fragment = importer.import_blob(blob.clone()).expect("import JSON");
+                    std::hint::black_box(fragment.facts().len());
                 });
             },
         );